backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.0
pydantic[email]==2.6.1
pydantic-settings==2.2.1
websockets==12.0
python-dateutil==2.8.2
email-validator==2.1.0.post1
